Refactor console output and create separate component
